+++ src/evaluation/evaluation.controller.ts
@@ -1,14 +1,18 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete, ValidationPipe, ParseIntPipe } from '@nestjs/common';
+import { Controller, Get, Post, Body, Patch, Param, Delete, ValidationPipe, ParseIntPipe, UnauthorizedException } from '@nestjs/common';
 import { EvaluationService } from './evaluation.service';
 import { CreateEvaluationDto } from './dto/createEvaluation.dto';
 import { UpdateEvaluationDto } from './dto/updateEvaluation.dto';
+import { CurrentUser } from 'src/auth/decorators/current-user.decorator';
+import { UserPayload } from 'src/auth/types/UserPayload';
 
 @Controller('evaluation')
 export class EvaluationController {
   constructor(private readonly evaluationService: EvaluationService) {}
 
   @Post()
-  create(@Body(ValidationPipe) EvaluationData: CreateEvaluationDto) {
+  create(@Body(ValidationPipe) EvaluationData: CreateEvaluationDto, @CurrentUser() currentUser: UserPayload) {
+    if (EvaluationData.userID !== currentUser.sub)
+      throw new UnauthorizedException("Só é possível criar avaliações para própria conta.")
     return this.evaluationService.create(EvaluationData);
   }
 
@@ -23,12 +27,19 @@ export class EvaluationController {
   }
 
   @Patch(':id')
-  update(@Param('id', ParseIntPipe) id: number, @Body(ValidationPipe) evaluationData: UpdateEvaluationDto) {
+  update(@Param('id', ParseIntPipe) id: number, @Body(ValidationPipe) evaluationData: UpdateEvaluationDto,@CurrentUser() currentUser: UserPayload) {
+    if (evaluationData.userID !== currentUser.sub){
+      throw new UnauthorizedException("Só é possível editar avaliações da própria conta.")
+    }
     return this.evaluationService.update(id, evaluationData);
   }
 
   @Delete(':id')
-  remove(@Param('id', ParseIntPipe) id: number) {
+  async remove(@Param('id', ParseIntPipe) id: number, @CurrentUser() currentUser: UserPayload) {
+    const evaluation= await this.evaluationService.findOne(id);
+    if (evaluation.userID !== currentUser.sub){
+      throw new UnauthorizedException("Só é possível editar avaliações da própria conta.")
+    }
     return this.evaluationService.remove(id);
   }
 }
